WIP Game of life
