+++ showml/simulations/conways_game_of_life/event.py
@@ -1,0 +1,11 @@
+from abc import ABC, abstractmethod
+
+class Event(ABC):
+    @abstractmethod
+    def run_event():
+        pass
+
+class StartEvent(Event):
+    def run_event():
+
+
